pipeline/pyproject.toml

[build-system]
requires = ["setuptools>=42", "wheel"]  # Build-time dependencies
build-backend = "setuptools.build_meta"

[project]
name = "museum_data_pipeline"
version = "1.0"
authors = [
{name = "Samuel Stern"}
]
dependencies = [
"appnope==0.1.4",
"astroid==3.3.8",
"asttokens==3.0.0",
"boto3==1.35.92",
"botocore==1.35.92",
"comm==0.2.2",
"confluent-kafka==2.8.0",
"contourpy==1.3.1",
"coverage==7.6.10",
"cycler==0.12.1",
"DateTime==5.5",
"debugpy==1.8.11",
"decorator==5.1.1",
"dill==0.3.9",
"executing==2.1.0",
"fonttools==4.55.3",
"iniconfig==2.0.0",
"ipykernel==6.29.5",
"ipython==8.31.0",
"isort==5.13.2",
"jedi==0.19.2",
"jmespath==1.0.1",
"jupyter_client==8.6.3",
"jupyter_core==5.7.2",
"kiwisolver==1.4.8",
"logging==0.4.9.6",
"matplotlib==3.10.0",
"matplotlib-inline==0.1.7",
"mccabe==0.7.0",
"nest-asyncio==1.6.0",
"numpy==2.2.1",
"packaging==24.2",
"parso==0.8.4",
"pexpect==4.9.0",
"pillow==11.1.0",
"platformdirs==4.3.6",
"pluggy==1.5.0",
"prompt_toolkit==3.0.48",
"psutil==6.1.1",
"psycopg2-binary==2.9.10",
"ptyprocess==0.7.0",
"pure_eval==0.2.3",
"Pygments==2.19.1",
"pylint==3.3.3",
"pyparsing==3.2.1",
"pytest==8.3.4",
"pytest-cov==6.0.0",
"python-dateutil==2.9.0.post0",
"python-dotenv==1.0.1",
"pytz==2024.2",
"pyzmq==26.2.0",
"s3transfer==0.10.4",
"setuptools==75.7.0",
"six==1.17.0",
"stack-data==0.6.3",
"tomlkit==0.13.2",
"tornado==6.4.2",
"traitlets==5.14.3",
"urllib3==2.3.0",
"wcwidth==0.2.13",
"zope.interface==7.2",
]

[project.urls]
Homepage = "https://github.com/stern-sigma"
